+++ packages/fluent_localization/fluent_localization/CHANGELOG.md
@@ -62,3 +62,8 @@
 ## 1.3.0
 
 * Upgrade fluent_sdk version to v0.4.0
+
+## 1.4.0
+
+* BREAKING CHANGE: Updated `fluent_sdk` dependency to `^0.5.0`
+* REFACTOR: Renamed `build` method to `onCreate` in `LocalizationModule`